% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_events.R
\name{create_events_df}
\alias{create_events_df}
\title{Create a dataframe of public events from GitHub}
\usage{
create_events_df()
}
\value{
dataframe (2D list)
}
\description{
Returns a dataframe with selected information about the 30 most recent events from Github. Information includes event type, actor, repo and payload if the event was a push.
}
\examples{
create_events_df()
}
