% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_events.R
\name{create_repo_df}
\alias{create_repo_df}
\title{Create a dataframe of events from a GitHub repository}
\usage{
create_repo_df(owner, repo)
}
\arguments{
\item{owner}{string}

\item{repo}{string}
}
\value{
dataframe (2D list)
}
\description{
Returns a dataframe with selected information about the 30 most recent events from a public repository (within the last 90 days). Information includes event id, event type, date, time, author and push message if the event was a push.
}
\examples{
create_repo_df("nganlyle", "534_project")
}
