% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_events.R
\name{get_events_lists}
\alias{get_events_lists}
\title{Create a table for actor and repo}
\usage{
get_events_lists()
}
\value{
table
}
\description{
Create a table for actor and repo
}
\examples{
get_events_lists()
}
