% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_events.R
\name{get_events_plots}
\alias{get_events_plots}
\title{Create plots for event type and payload size}
\usage{
get_events_plots()
}
\value{
plot
}
\description{
Create plots for event type and payload size
}
\examples{
get_events_plots()
}
