% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_events.R
\name{get_repo_events}
\alias{get_repo_events}
\title{Get a list of events from a GitHub respository}
\usage{
get_repo_events(owner, repo)
}
\arguments{
\item{owner}{string}

\item{repo}{string}
}
\value{
list
}
\description{
Returns the 30 most recent events from a public repository (within the last 90 days), including all metadata, as a list of lists.
}
\examples{
get_repo_events("nganlyle", "534_project")
}
