% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_events.R
\name{get_repo_log}
\alias{get_repo_log}
\title{Create events stamp for a repository}
\usage{
get_repo_log(owner, repo, log_date)
}
\arguments{
\item{owner}{string}

\item{repo}{string}

\item{log_date}{string yyyy-mm-dd}
}
\value{
dataframe (2D list)
}
\description{
Create events stamp for a repository
}
\examples{
get_repo_log("nganlyle", "534_project", "2020-01-26")
}
